requirements.txt
discord.py==2.5.2
google-genai==1.11.0
python-dotenv==1.1.0